+++ apps/desktop/src/lib/trpc/routers/workspaces/utils/setup.ts
@@ -4,7 +4,11 @@ import { CONFIG_FILE_NAME, PROJECT_SUPERSET_DIR_NAME } from "shared/constants";
 import type { SetupConfig } from "shared/types";
 
 export function loadSetupConfig(mainRepoPath: string): SetupConfig | null {
-	const configPath = join(mainRepoPath, PROJECT_SUPERSET_DIR_NAME, CONFIG_FILE_NAME);
+	const configPath = join(
+		mainRepoPath,
+		PROJECT_SUPERSET_DIR_NAME,
+		CONFIG_FILE_NAME,
+	);
 
 	if (!existsSync(configPath)) {
 		return null;
