+++ apps/desktop/src/renderer/components/ConfigFilePreview/index.ts
@@ -1,2 +1,4 @@
-export { ConfigFilePreview, type ConfigFilePreviewProps } from "./ConfigFilePreview";
-
+export {
+	ConfigFilePreview,
+	type ConfigFilePreviewProps,
+} from "./ConfigFilePreview";
