+++ apps/desktop/src/renderer/components/SetupConfigModal/SetupConfigModal.tsx
@@ -1,4 +1,3 @@
-import { Button } from "@superset/ui/button";
 import {
 	Dialog,
 	DialogContent,
@@ -6,7 +5,6 @@ import {
 	DialogHeader,
 	DialogTitle,
 } from "@superset/ui/dialog";
-import { HiArrowTopRightOnSquare } from "react-icons/hi2";
 import { ConfigFilePreview } from "renderer/components/ConfigFilePreview";
 import { trpc } from "renderer/lib/trpc";
 import {
@@ -14,7 +12,7 @@ import {
 	useConfigModalOpen,
 	useConfigModalProjectId,
 } from "renderer/stores/config-modal";
-import { CONFIG_FILE_NAME, WEBSITE_URL } from "shared/constants";
+import { CONFIG_FILE_NAME } from "shared/constants";
 
 export function SetupConfigModal() {
 	const isOpen = useConfigModalOpen();
@@ -33,18 +31,14 @@ export function SetupConfigModal() {
 
 	const projectName = project?.name ?? "your-project";
 
-	const handleLearnMore = () => {
-		window.open(`${WEBSITE_URL}/scripts`, "_blank");
-	};
-
 	return (
 		<Dialog modal open={isOpen} onOpenChange={(open) => !open && closeModal()}>
 			<DialogContent className="sm:max-w-2xl">
 				<DialogHeader>
 					<DialogTitle>Configure scripts</DialogTitle>
 					<DialogDescription>
-						Edit {CONFIG_FILE_NAME} to automate setting up workspaces and running your
-						app.
+						Edit {CONFIG_FILE_NAME} to automate setting up workspaces and
+						running your app.
 					</DialogDescription>
 				</DialogHeader>
 
@@ -53,18 +47,6 @@ export function SetupConfigModal() {
 					configFilePath={configFilePath ?? undefined}
 					className="mt-4"
 				/>
-
-				<div className="mt-4">
-					<Button
-						variant="outline"
-						size="sm"
-						onClick={handleLearnMore}
-						className="gap-2"
-					>
-						Learn how to use scripts
-						<HiArrowTopRightOnSquare className="h-4 w-4" />
-					</Button>
-				</div>
 			</DialogContent>
 		</Dialog>
 	);
